requirements.txt
streamlit==1.1.0
pandas==1.3.3
pandas-datareader==0.10.0
pandocfilters==1.5.0
numpy==1.21.2
requests==2.26.0
requests-html==0.10.0
requests-oauthlib==1.3.0
tweepy==3.9.0
config==0.5.1
postgres==4.0
psycopg2-binary==2.9.1
plotly==5.3.1
yfinance==0.1.63
cufflinks==0.17.3


